Add 'no resource found' message and move generateRandomLink to links
